Assign juntas vecinales and neighborhoods correctly
Clean up controller logic and push it into the controller.

Assigning sector to a user when it's already reachable through the colonium association is kind of odd, but instead of removing it, I ensured it was always assigned correctly as a model callback. The reason I went against removing it is because we have broken tests and currently there is no way of being safe that this change will not have any negative side effects.